lxd/events: Move common functionality between external and devlxd servers into common file and types
